etl_pricedumping/requirements.txt
python-dotenv
requests
web3
uvicorn
fastapi
